Move splitter item serialisation utils from config_layout.cpp
