+++ foo_ui_columns/foo_ui_columns.vcxproj
@@ -446,6 +446,7 @@
     <ClCompile Include="list_view_panel.cpp" />
     <ClCompile Include="notification_area.cpp" />
     <ClCompile Include="rebar_band.cpp" />
+    <ClCompile Include="splitter_utils.cpp" />
     <ClCompile Include="tab_pview_artwork.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -508,6 +509,7 @@
     <ClInclude Include="list_view_panel.h" />
     <ClInclude Include="notification_area.h" />
     <ClInclude Include="rebar_band.h" />
+    <ClInclude Include="splitter_utils.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include=".clang-format" />
