+++ foo_ui_columns/foo_ui_columns.vcxproj.filters
@@ -463,6 +463,9 @@
     <ClCompile Include="tab_pview_artwork.cpp">
       <Filter>Prefs UI</Filter>
     </ClCompile>
+    <ClCompile Include="splitter_utils.cpp">
+      <Filter>Helpers</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include=".\cache.h">
@@ -639,6 +642,9 @@
     </ClInclude>
     <ClInclude Include="notification_area.h">
       <Filter>Main Window</Filter>
+    </ClInclude>
+    <ClInclude Include="splitter_utils.h">
+      <Filter>Helpers</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
